olim/requirements.txt
Django
django-cors-headers
djangorestframework
sqlparse
gunicorn